refactor(wl): split overlay rendering modules
Break the monolithic overlay module into focused files without changing overlay behavior:
- Keep `overlay/mod.rs` as a small barrel and HUD orchestrator.
- Move shared overlay chip rendering into `overlay/chip.rs`.
- Move action-row/keycap layout and drawing into `overlay/action_row.rs`.
- Move shared overlay text clipping/truncation helpers into `overlay/text.rs`.
- Move cluster overflow strip rendering, promotion animation, drag preview drawing, and hit-testing into `overlay/cluster_overflow.rs`.
- Move persistent mode banner drawing into `overlay/banner.rs`.
- Move toast/error-toast layout, drawing, scrolling, and hit-testing into `overlay/toast.rs`.
- Move focus-cycle switcher rendering into `overlay/focus_cycle.rs`.
- Move exit confirmation rendering into `overlay/exit_confirm.rs`.
- Move hover-label rendering into `overlay/hover_label.rs`.
- Move cluster selection marker rendering into `overlay/selection_marker.rs`.
- Preserve the existing `crate::overlay::*` public API by re-exporting the same externally used functions and types.

This is a mechanical maintainability refactor. It keeps rendering, input hit-testing, overlay styling, screenshots, cluster naming, and test behavior unchanged while making future overlay changes easier to review.
